pyproject.toml
[project]
name = "web-status"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "telegrinder>=0.5.1",
]
